# 8275646: Implement optimized upcall stubs on AArch64