--- a/src/hotspot/share/prims/universalUpcallHandler.cpp
+++ b/src/hotspot/share/prims/universalUpcallHandler.cpp
@@ -103,8 +103,6 @@ JavaThread* ProgrammableUpcallHandler::on_entry(OptimizedEntryBlob::FrameData* c
   debug_only(thread->inc_java_call_counter());
   thread->set_active_handles(context->new_handles);     // install new handle block and reset Java frame linkage
 
-  MACOS_AARCH64_ONLY(thread->enable_wx(WXExec));
-
   return thread;
 }
 
@@ -112,8 +110,6 @@ JavaThread* ProgrammableUpcallHandler::on_entry(OptimizedEntryBlob::FrameData* c
 void ProgrammableUpcallHandler::on_exit(OptimizedEntryBlob::FrameData* context) {
   JavaThread* thread = context->thread;
   assert(thread == JavaThread::current(), "must still be the same thread");
-
-  MACOS_AARCH64_ONLY(thread->enable_wx(WXWrite));
 
   // restore previous handle block
   thread->set_active_handles(context->old_handles);