@theRealAph 2021-11-05
The AArch64 code parts look reasonable enough, but parts of them seem to repeat some logic already present in MacroAssembler. I guess as it's desirable to stay close to the x86 port that's OK.
I'm interested to know where all that 250ns is going. Did you look at -prof perfasm?