@theRealAph 2021-11-08
> > I removed the calls to change the W^X state in on_entry/on_exit calls:
> > in the on_entry case the stub must already be executable because we
> > called into the VM from there, and for on_exit we need the code to be
> > executable not writable otherwise we'll get a SIGBUS as soon as we
> > return to the stub.
> 
> Thanks for cleaning this up as well. That code was taken from `JavaCallWrapper` code. Do you happen to know which use-case those calls were supposed to address? (I'm assuming things still work on MacOS/AArch64 without them).

This is in case native calls cause JIT patching.