@theRealAph 2021-11-09
> > Tried on MacOS, this is _very_ weird:
> > ```
> > ....[Hottest Methods (after inlining)]..............................................................
> >  25.52%               runtime stub  StubRoutines::updateBytesAdler32 
> > ```
> 
> I think it's a bug in JMH's parsing of the `-XX:+PrintStubCode` output: it's expecting these to all be prefixed with `StubRoutines::` but many aren't including the `optimized_upcall_stub` added by this patch. `StubRoutines::updateBytesAdler32` is the last one printed with that prefix on MacOS, and likewise `StubRoutines::atomic entry points` is the last one printed on Linux (allegedly the third hottest method above). So I think it's been misattributed. I'll see if I can fix it...

That makes perfect sense. I'm making rather a fuss about this because I think I know how long a direct call _should_ make_, and indeed does make in C++. It could be that this really is the best we can do.